trmenu/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-js-asset==2.2.0
flake8==7.0.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
isort==5.13.2
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
pep8-naming==0.13.3
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
